pyproject.toml
[project]
name = "edr-capstone"
version = "0.1.0"
description = "Engineering Document Review — Multi-Agent Compliance Platform"
requires-python = ">=3.11"
dependencies = [
    # LLM & Agent Framework
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    # Vector Store
    "langchain-qdrant>=0.2.0",
    "qdrant-client>=1.12.0",
    # PDF Loading
    "pypdf>=4.0.0",
    # Advanced Pipeline — Azure Document Intelligence + Hybrid Retrieval
    "azure-ai-documentintelligence>=1.0.0",
    "fastembed>=0.4.0",
    # Search
    "tavily-python>=0.5.0",
    "langchain-tavily>=0.1.0",
    # Monitoring
    "langsmith>=0.1.0",
    # UI
    "streamlit>=1.38.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    # Evaluation
    "ragas>=0.4.0",
    "tabulate>=0.9.0",
    "openevals>=0.1.0",
    "nltk>=3.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src", "evaluation"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
